Backend/orders-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-dotenv
python-jose[cryptography]
